Break out matrix multiplication code in ntt.rs.
